Update package structure and uplift dependencies
