+++ conf/application.conf
@@ -19,36 +19,15 @@ include "backend.conf"
 
 appName = one-stop-shop-returns
 
-# An ApplicationLoader that uses Guice to bootstrap the application.
 play.application.loader = "uk.gov.hmrc.play.bootstrap.ApplicationLoader"
-
-# Primary entry point for all HTTP requests on Play applications
 play.http.requestHandler = "uk.gov.hmrc.play.bootstrap.http.RequestHandler"
-
-# Provides an implementation of AuditConnector. Use `uk.gov.hmrc.play.bootstrap.AuditModule` or create your own.
-# An audit connector must be provided.
 play.modules.enabled += "uk.gov.hmrc.play.bootstrap.AuditModule"
-
-# Provides an implementation of MetricsFilter. Use `uk.gov.hmrc.play.graphite.GraphiteMetricsModule` or create your own.
-# A metric filter must be provided
 play.modules.enabled += "uk.gov.hmrc.play.bootstrap.graphite.GraphiteMetricsModule"
-
-# Provides an implementation and configures all filters required by a Platform frontend microservice.
 play.modules.enabled += "uk.gov.hmrc.play.bootstrap.backend.BackendModule"
-
-# Default http client
 play.modules.enabled += "uk.gov.hmrc.play.bootstrap.HttpClientModule"
 
-# Json error handler
 play.http.errorHandler = "uk.gov.hmrc.play.bootstrap.backend.http.JsonErrorHandler"
 
-# Play Modules
-# ~~~~
-# Additional play modules can be added here
-
-
-# The application languages
-# ~~~~~
 play.i18n.langs = ["en"]
 
 # Router
@@ -63,24 +42,6 @@ play.i18n.langs = ["en"]
 # !!!WARNING!!! DO NOT CHANGE THIS ROUTER
 play.http.router = prod.Routes
 
-# Controller
-# ~~~~~
-# By default all controllers will have authorisation, logging and
-# auditing (transaction monitoring) enabled.
-# The below controllers are the default exceptions to this rule.
-
-controllers {
-  # 300 is the default, you may need to change this according to your needs
-  confidenceLevel = 300
-
-  uk.gov.hmrc.onestopshopreturns.controllers.MicroserviceHelloWorldController = {
-    needsAuth = false
-    needsLogging = false
-    needsAuditing = false
-  }
-}
-
-# Metrics plugin settings - graphite reporting is configured on a per env basis
 metrics {
   name = ${appName}
   rateUnit = SECONDS
@@ -121,4 +82,3 @@ microservice {
     }
   }
 }
-
