+++ project/plugins.sbt
@@ -2,8 +2,8 @@ resolvers += "HMRC-open-artefacts-maven" at "https://open.artefacts.tax.service.
 resolvers += Resolver.url("HMRC-open-artefacts-ivy", url("https://open.artefacts.tax.service.gov.uk/ivy2"))(Resolver.ivyStylePatterns)
 resolvers += Resolver.typesafeRepo("releases")
 
-addSbtPlugin("uk.gov.hmrc"       % "sbt-auto-build"     % "3.2.0")
-addSbtPlugin("uk.gov.hmrc"       % "sbt-git-versioning" % "2.3.0")
+addSbtPlugin("uk.gov.hmrc"       % "sbt-auto-build"     % "3.3.0")
+addSbtPlugin("uk.gov.hmrc"       % "sbt-git-versioning" % "2.4.0")
 addSbtPlugin("uk.gov.hmrc"       % "sbt-distributables" % "2.1.0")
 addSbtPlugin("com.typesafe.play" % "sbt-plugin"         % "2.8.7")
