+++ test/controllers/MicroserviceHelloWorldControllerSpec.scala
@@ -14,17 +14,23 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.onestopshopreturns.config
+package controllers
 
-import javax.inject.{Inject, Singleton}
-import play.api.Configuration
-import uk.gov.hmrc.play.bootstrap.config.ServicesConfig
+import org.scalatest.matchers.should.Matchers
+import org.scalatest.wordspec.AnyWordSpec
+import play.api.http.Status
+import play.api.test.Helpers._
+import play.api.test.{FakeRequest, Helpers}
 
-@Singleton
-class AppConfig @Inject()
-  (
-    config: Configuration
-  , servicesConfig: ServicesConfig
-  ) {
+class MicroserviceHelloWorldControllerSpec extends AnyWordSpec with Matchers {
 
+  private val fakeRequest = FakeRequest("GET", "/")
+  private val controller = new MicroserviceHelloWorldController(Helpers.stubControllerComponents())
+
+  "GET /" should {
+    "return 200" in {
+      val result = controller.hello()(fakeRequest)
+      status(result) shouldBe Status.OK
+    }
+  }
 }
